Extract lightweight catalog and cruises contract packages
Add @voyantjs/catalog-contracts and @voyantjs/cruises-contracts as
zod-only packages that own the pure adapter and content contracts, so
external consumers can validate against the framework without the
catalog/cruises runtime. The runtime packages re-export from them to
keep existing import paths stable.

- Carry the cruise cabin feature/bed/accessibility/view-type facets in
  the contract schema; the facet vocabularies live in cruises-contracts
  and are re-exported from cruises.
- Make the catalog content service re-export the pure content primitives
  (isStale, pickBestCachedLocale, JSON-pointer overlays,
  mergeOverlaysIntoContent) from catalog-contracts, keeping only the
  runtime-bound (Drizzle/Postgres) primitives as a single source of truth.

Resolves #1400.
