+++ .changeset/light-contract-packages.md
@@ -1,0 +1,24 @@
+---
+"@voyantjs/catalog-contracts": minor
+"@voyantjs/cruises-contracts": minor
+"@voyantjs/catalog": minor
+"@voyantjs/cruises": minor
+---
+
+Add lightweight catalog and cruises contract packages for external consumers.
+
+`@voyantjs/catalog-contracts` now owns the pure catalog adapter contracts,
+adapter Zod schemas, field-policy contracts, provenance, drift event payloads,
+and pure content locale/overlay helpers. `@voyantjs/cruises-contracts` now owns
+the `cruises/v1` rich content schema (including the cabin feature, bed,
+accessibility, and view-type facet vocabularies), version, types, and validator.
+
+The pure content primitives (`isStale`, `pickBestCachedLocale`, the JSON-pointer
+overlay applier, and `mergeOverlaysIntoContent`) now have a single source of
+truth in `@voyantjs/catalog-contracts`; `@voyantjs/catalog`'s content service
+re-exports them and retains only the runtime-bound (Drizzle/Postgres) primitives.
+The cruise cabin facet vocabularies likewise live in `@voyantjs/cruises-contracts`
+and are re-exported from `@voyantjs/cruises`.
+
+The existing `@voyantjs/catalog` and `@voyantjs/cruises` contract import paths
+remain available through compatibility re-exports.
